+++ DevotionTweaks/Devotion/Components/BetterLemurController.cs
@@ -1,46 +1,90 @@
 ﻿using LemurFusion;
 using LemurFusion.Config;
 using RoR2;
-using System.Collections.Generic;
-using System.Runtime.CompilerServices;
 using LemurFusion.Devotion;
 using LemurFusion.Devotion.Components;
 using UnityEngine.AddressableAssets;
 using UnityEngine;
+using System.Linq;
+using System.Collections.Generic;
+using System.Runtime.CompilerServices;
 
 public class BetterLemurController : DevotedLemurianController
 {
     #region Lemur Instance
-    public SortedList<ItemIndex, int> _devotedItemList { get; set; } = [];
-    public SortedList<ItemIndex, int> _untrackedItemList { get; set; } = [];
-    
     public BetterInventoryController BetterInventoryController => base._devotionInventoryController as BetterInventoryController;
     public new Inventory LemurianInventory => this._lemurianMaster ? this._lemurianMaster.inventory : null;
+    public Inventory PersonalInventory { get; set; }
+
+    public void InitializeDevotedLemurian()
+    {
+        this._leashDistSq = PluginConfig.teleportDistance.Value * PluginConfig.teleportDistance.Value;
+
+        if (!this.PersonalInventory)
+            this.PersonalInventory = this._lemurianMaster.GetComponents<Inventory>().Last();
+        this.PersonalInventory.GiveItem(this.DevotionItem);
+
+        ShareItem(this.DevotionItem);
+
+        var inv = this._lemurianMaster.inventory;
+        inv.GiveItem(this.DevotionItem);
+        inv.GiveItem(CU8Content.Items.LemurianHarness.itemIndex);
+        inv.ResetItem(RoR2Content.Items.MinionLeash.itemIndex, 1);
+        inv.ResetItem(RoR2Content.Items.UseAmbientLevel.itemIndex, 1);
+        inv.ResetItem(RoR2Content.Items.TeleportWhenOob.itemIndex, 1);
+        if (LemurFusionPlugin.riskyInstalled)
+            AddRiskyAllyItem();
+    }
+
+    public void ShareItem(ItemIndex item)
+    {
+        if (PluginConfig.enableSharedInventory.Value)
+        {
+            if (LemurianInventory.GetItemCount(CU8Content.Items.LemurianHarness) == 0)
+                LemurianInventory.AddItemsFrom(this.BetterInventoryController._devotionMinionInventory, ConfigExtended.Blacklist_Filter);
+
+            foreach (var friend in this.BetterInventoryController.GetFriends(this))
+            {
+                friend.LemurianInventory.GiveItem(item);
+            }
+        }
+    }
 
     [MethodImpl(MethodImplOptions.NoInlining | MethodImplOptions.NoOptimization)]
     public void AddRiskyAllyItem()
     {
-        var item = RiskyMod.Allies.AllyItems.AllyMarkerItem.itemIndex;
-        Utils.SetItem(this._untrackedItemList, item);
+        LemurianInventory.ResetItem(RiskyMod.Allies.AllyItems.AllyMarkerItem.itemIndex, 1);
     }
 
     public void KillYourSelf()
     {
-        if (DevotionTweaks.instance.EnableSharedInventory && this.BetterInventoryController)
-            this.BetterInventoryController.RemoveSharedItemsFromFriends(this._devotedItemList);
+        bool sharing = DevotionTweaks.instance.EnableSharedInventory && this.BetterInventoryController;
+        var friends = sharing ? this.BetterInventoryController.GetFriends() : [];
 
-        var dropType = ConfigExtended.DeathDrop_ItemType.Value;
-        if (dropType != DevotionTweaks.DeathItem.None && this.LemurianBody)
+        if (this.PersonalInventory)
         {
-            foreach (var item in this._devotedItemList)
+            foreach (var item in this.PersonalInventory.itemAcquisitionOrder.ToList())
             {
-                var pickupIndex = FindPickupIndex(item.Key, dropType);
+                var count = this.PersonalInventory.GetItemCount(item);
+                if (count <= 0)
+                    continue;
+
+                if (sharing)
+                {
+                    this.BetterInventoryController.RemoveItem(item, System.Math.Min(count, this.BetterInventoryController._devotionMinionInventory.GetItemCount(item)));
+                    foreach (var friend in friends)
+                    {
+                        friend.LemurianInventory.RemoveItem(item, System.Math.Min(count, friend.LemurianInventory.GetItemCount(item)));
+                    }
+                }
+
+                var pickupIndex = FindPickupIndex(item);
                 if (pickupIndex != PickupIndex.none)
                 {
-                    var dropCount = ConfigExtended.DeathDrop_DropAll.Value ? item.Value : 1;
+                    var dropCount = ConfigExtended.DeathDrop_DropAll.Value ? count : 1;
                     for (var i = 0; i < dropCount; i++)
                     {
-                        PickupDropletController.CreatePickupDroplet(pickupIndex, this.LemurianBody.corePosition, UnityEngine.Random.insideUnitCircle * 15f);
+                        PickupDropletController.CreatePickupDroplet(pickupIndex, this.LemurianBody.corePosition, Random.insideUnitCircle * 15f);
                     }
                 }
             }
@@ -62,30 +106,29 @@ public class BetterLemurController : DevotedLemurianController
             GameObject.Destroy(this._lemurianMaster.gameObject, 1f);
     }
 
-    public PickupIndex FindPickupIndex(ItemIndex itemIndex, DevotionTweaks.DeathItem dropType)
+    public PickupIndex FindPickupIndex(ItemIndex itemIndex)
     {
-        if (itemIndex != ItemIndex.None)
+        if (itemIndex != ItemIndex.None && this.LemurianBody)
         {
-            switch (dropType)
+            switch (ConfigExtended.DeathDrop_ItemType.Value)
             {
                 // this is a warcrime
                 case DevotionTweaks.DeathItem.Scrap:
+                    return ItemCatalog.GetItemDef(itemIndex).tier switch
                     {
-                        return ItemCatalog.GetItemDef(itemIndex).tier switch
-                        {
-                            ItemTier.Tier1 => PickupCatalog.FindPickupIndex("ItemIndex.ScrapWhite"),
-                            ItemTier.Tier2 => PickupCatalog.FindPickupIndex("ItemIndex.ScrapGreen"),
-                            ItemTier.Tier3 => PickupCatalog.FindPickupIndex("ItemIndex.ScrapRed"),
-                            ItemTier.Boss => PickupCatalog.FindPickupIndex("ItemIndex.ScrapYellow"),
-                            _ => PickupIndex.none,
-                        };
-                    }
+                        ItemTier.Tier1 => PickupCatalog.FindPickupIndex("ItemIndex.ScrapWhite"),
+                        ItemTier.Tier2 => PickupCatalog.FindPickupIndex("ItemIndex.ScrapGreen"),
+                        ItemTier.Tier3 => PickupCatalog.FindPickupIndex("ItemIndex.ScrapRed"),
+                        ItemTier.Boss => PickupCatalog.FindPickupIndex("ItemIndex.ScrapYellow"),
+                        _ => PickupIndex.none,
+                    };
 
                 case DevotionTweaks.DeathItem.Original:
                     return PickupCatalog.FindPickupIndex(itemIndex);
 
                 case DevotionTweaks.DeathItem.Custom:
-                    if (ConfigExtended.DeathDrops_TierToItem_Map.TryGetValue(ItemCatalog.GetItemDef(itemIndex).tier, out var idx) && idx != ItemIndex.None)
+                    var itemDef = ItemCatalog.GetItemDef(itemIndex);
+                    if (itemDef && ConfigExtended.DeathDrops_TierToItem_Map.TryGetValue(itemDef.tier, out var idx) && idx != ItemIndex.None)
                         return PickupCatalog.FindPickupIndex(idx);
                     break;
             }