+++ DevotionTweaks/Devotion/StatTweaks.cs
@@ -4,6 +4,8 @@ using UnityEngine.Networking;
 using UnityEngine;
 using RoR2;
 using LemurFusion.Config;
+using System.Linq;
+using static RoR2.FriendlyFireManager;
 
 namespace LemurFusion.Devotion
 {
@@ -29,14 +31,7 @@ namespace LemurFusion.Devotion
 
         #region Hooks
         #region UI
-        private static string Util_GetBestMasterName(On.RoR2.Util.orig_GetBestMasterName orig, CharacterMaster characterMaster)
-        {
-            if (Utils.IsDevoted(characterMaster))
-            {
-                return characterMaster.GetBody()?.GetDisplayName();
-            }
-            return orig(characterMaster);
-        }
+        private static string Util_GetBestMasterName(On.RoR2.Util.orig_GetBestMasterName orig, CharacterMaster characterMaster) => Utils.IsDevoted(characterMaster) ? characterMaster.GetBody()?.GetDisplayName() : orig(characterMaster);
 
         private static string CharacterBody_GetDisplayName(On.RoR2.CharacterBody.orig_GetDisplayName orig, CharacterBody self)
         {
@@ -57,88 +52,83 @@ namespace LemurFusion.Devotion
         {
             orig(self);
             if (!PluginConfig.enableMinionScoreboard.Value)
-            {
                 return;
-            }
 
-            List<CharacterMaster> masters = [];
-            foreach (var instance in PlayerCharacterMasterController.instances)
+            List<BetterLemurController> lems = [];
+            foreach (var instance in PlayerCharacterMasterController.instances.Where(m => m.hasEffectiveAuthority))
             {
-                masters.Add(instance.master);
-                var minionGroup = MinionOwnership.MinionGroup.FindGroup(instance.master.netId);
+                var minionGroup = MinionOwnership.MinionGroup.FindGroup(instance.netId);
                 if (minionGroup != null)
                 {
                     foreach (var minionOwnership in minionGroup.members)
                     {
                         if (minionOwnership && minionOwnership.TryGetComponent<BetterLemurController>(out var lemCtrl) && lemCtrl.LemurianInventory)
                         {
-                            masters.Add(lemCtrl._lemurianMaster);
+                            lems.Add(lemCtrl);
                         }
                     }
                 }
             }
 
-            self.SetStripCount(masters.Count);
-            for (var i = 0; i < masters.Count; i++)
+            if (lems.Any())
             {
-                self.stripAllocator.elements[i].SetMaster(masters[i]);
-            }
-            /*
-            self.SetStripCount(masters.Count + lemCtrlList.Count);
-            for (var i = 0; i < masters.Count; i++)
-            {
-                self.stripAllocator.elements[i].SetMaster(masters[i]);
-            }
-            for (var i = 0; i < lemCtrlList.Count; i++)
-            {
-                var lemCtrl = lemCtrlList[i];
-                var strip = self.stripAllocator.elements[i + masters.Count];
-                strip.SetMaster(lemCtrl._lemurianMaster);
-
-                //lazy
-                if (!PluginConfig.showPersonalInventory.Value)
-                    break;
-
-                lemCtrl.LemurianInventory.onInventoryChanged -= strip.itemInventoryDisplay.OnInventoryChanged;
-
-                strip.itemInventoryDisplay.itemOrderCount = lemCtrl._devotedItemList.Count;
-                Array.Copy(lemCtrl.LemurianInventory.itemStacks, strip.itemInventoryDisplay.itemStacks, strip.itemInventoryDisplay.itemStacks.Length);
-                Array.Copy(lemCtrl._devotedItemList.Keys.ToArray(), strip.itemInventoryDisplay.itemOrder, strip.itemInventoryDisplay.itemOrderCount);
-                foreach (var item in lemCtrl._devotedItemList)
+                int start = self.stripAllocator.elements.Count;
+                self.SetStripCount(start + lems.Count);
+                for (int i = 0; i < lems.Count; i++)
                 {
-                    strip.itemInventoryDisplay.itemStacks[(int)item.Key] = item.Value;
+                    self.stripAllocator.elements[start + i].SetMaster(lems[i]._lemurianMaster);
+                    if (PluginConfig.showPersonalInventory.Value)
+                        self.stripAllocator.elements[start + i].itemInventoryDisplay.SetSubscribedInventory(lems[i].PersonalInventory);
                 }
-                strip.itemInventoryDisplay.RequestUpdateDisplay();
-            }*/
+            }
         }
         #endregion
 
         #region Stats
-        private static void CharacterBody_onBodyStartGlobal(CharacterBody body)
+        private static void CharacterBody_onBodyStartGlobal(CharacterBody lemBody)
         {
-            if (Utils.IsDevoted(body))
+            if (Utils.IsDevoted(lemBody))
             {
                 if (AITweaks.disableFallDamage.Value)
-                    body.bodyFlags |= CharacterBody.BodyFlags.IgnoreFallDamage;
+                    lemBody.bodyFlags |= CharacterBody.BodyFlags.IgnoreFallDamage;
                 if (AITweaks.immuneToVoidDeath.Value)
-                    body.bodyFlags |= CharacterBody.BodyFlags.ImmuneToVoidDeath | CharacterBody.BodyFlags.ResistantToAOE;
-
-                if (body.modelLocator)
+                    lemBody.bodyFlags |= CharacterBody.BodyFlags.ImmuneToVoidDeath | CharacterBody.BodyFlags.ResistantToAOE;
+                
+                if (NetworkServer.active && lemBody.hurtBoxGroup && lemBody.hurtBoxGroup.hurtBoxes.Length > 0)
                 {
-                    var transform = body.modelLocator.modelTransform;
+                    var hurtBoxes = lemBody.hurtBoxGroup.hurtBoxes;
+                    foreach (var tc in TeamComponent.GetTeamMembers(TeamIndex.Player).Where(tc => tc && tc.body && tc.body.isPlayerControlled))
+                    {
+                        var teamBoxes = tc.body.hurtBoxGroup ? tc.body.hurtBoxGroup.hurtBoxes : null;
+                        if (teamBoxes?.Any() == true && FriendlyFireManager.friendlyFireMode == FriendlyFireMode.Off)
+                        {
+                            for (int i = 0; i < teamBoxes.Length; i++)
+                            {
+                                for (int j = 0; j < hurtBoxes.Length; j++)
+                                {
+                                    Physics.IgnoreCollision(teamBoxes[i].collider, hurtBoxes[j].collider, true);
+                                } // end for
+                            } // end for
+                        }
+                    }
+                }
+
+                if (NetworkClient.active && lemBody.modelLocator)
+                {
+                    var transform = lemBody.modelLocator.modelTransform;
                     if (transform)
                     {
-                        if (PluginConfig.miniElders.Value && body.bodyIndex == BodyCatalog.FindBodyIndex(DevotionTweaks.devotedBigLemBodyName) && NetworkClient.active)
+                        if (PluginConfig.miniElders.Value && lemBody.bodyIndex == BodyCatalog.FindBodyIndex(DevotionTweaks.devotedBigLemBodyName))
                         {
                             var init = PluginConfig.initScaleValue.Value * 0.01f;
                             var stack = PluginConfig.scaleValue.Value * 0.01f;
-                            var count = body.master.inventory.GetItemCount(CU8Content.Items.LemurianHarness);
+                            var count = lemBody.master.inventory.GetItemCount(CU8Content.Items.LemurianHarness);
 
                             transform.localScale = Vector3.one * (init + (stack * count));
                         }
 
                         if (transform.TryGetComponent<FootstepHandler>(out var footstep))
-                        footstep.baseFootstepString = "";
+                            footstep.baseFootstepString = "";
                     }
                 }
             }
@@ -146,14 +136,14 @@ namespace LemurFusion.Devotion
 
         private static void RecalculateStatsAPI_GetStatCoefficients(CharacterBody sender, RecalculateStatsAPI.StatHookEventArgs args)
         {
-            if (Utils.IsDevoted(sender) && sender.master.inventory && sender.master.TryGetComponent<BetterLemurController>(out var lem) && lem.LemurianInventory)
+            if (Utils.IsDevoted(sender) && sender.master.TryGetComponent<BetterLemurController>(out var lem) && lem.LemurianInventory)
             {
                 var fusionCount = lem.LemurianInventory.GetItemCount(CU8Content.Items.LemurianHarness);
                 if (PluginConfig.rebalanceHealthScaling.Value)
                 {
                     args.baseHealthAdd += sender.level * sender.levelMaxHealth * Utils.GetLevelModifier(lem.DevotedEvolutionLevel);
                     args.baseRegenAdd += sender.level * (sender.outOfCombat ? Utils.GetLevelModifier(lem.DevotedEvolutionLevel) : Utils.GetLevelModifier(0));
-                    args.armorAdd += 20f + (5f * Utils.GetLevelModifier(lem.DevotedEvolutionLevel));
+                    args.armorAdd += 20f + Utils.GetLevelModifier(lem.DevotedEvolutionLevel);
 
                     args.healthMultAdd += Utils.GetFusionStatMultiplier(PluginConfig.statMultHealth.Value, fusionCount, lem.DevotedEvolutionLevel);
                     args.damageMultAdd += Utils.GetFusionStatMultiplier(PluginConfig.statMultDamage.Value, fusionCount, lem.DevotedEvolutionLevel);